requirements.txt
aioredis==1.1.0
alabaster==0.7.12
aniso8601==3.0.2
arrow==0.12.1
asgiref==2.3.2
asn1crypto==0.24.0
async-timeout==3.0.0
atomicwrites==1.2.1
attrs==18.2.0
autobahn==18.9.1
Automat==0.7.0
Babel==2.6.0
backcall==0.1.0
certifi==2018.8.24
cffi==1.11.5
channels==2.1.3
channels-redis==2.3.0
chardet==3.0.4
CommonMark==0.5.4
constantly==15.1.0
contextlib2==0.5.5
coreapi==2.3.3
coreschema==0.0.4
coverage==4.5.1
cryptography==2.3.1
daphne==2.2.2
decorator==4.3.0
defusedxml==0.5.0
Django==2.1.1
django-allauth==0.37.1
django-cors-headers==2.4.0
django-crispy-forms==1.7.2
django-debug-toolbar==1.10.1
django-elasticsearch-dsl==0.5.1
django-elasticsearch-dsl-drf==0.17.1
django-extensions==2.1.2
django-extra-fields==1.0.0
django-model-utils==3.1.2
django-nine==0.1.13
django-phonenumber-field==2.0.1
django-pipeline==1.6.14
django-pipeline-browserify==0.6.2
django-rest-auth==0.9.3
django-rest-swagger==2.2.0
django-split-settings==0.3.0
django-taggit==0.23.0
django-taggit-serializer==0.1.7
django-taggit-templatetags==0.2.5
django-templatetag-sugar==1.0
django-webpack-loader==0.6.0
djangorestframework==3.8.2
docopt==0.6.2
docutils==0.14
elasticsearch==6.3.1
elasticsearch-dsl==6.1.0
epc==0.0.5
factory-boy==2.11.1
Faker==0.8.8
future==0.16.0
graphene==2.1.3
graphene-django==2.2.0
graphql-core==2.1
graphql-relay==0.4.5
gunicorn==19.9.0
h2==3.0.1
hiredis==0.2.0
hpack==3.0.0
hyperframe==5.1.0
hyperlink==18.0.0
idna==2.7
imagesize==1.1.0
incremental==17.5.0
ipaddress==1.0.22
ipdb==0.11
ipython==6.5.0
ipython-genutils==0.2.0
itypes==1.1.0
jedi==0.12.1
jediepcserver==0.2.7
Jinja2==2.10
jmespath==0.9.3
MarkupSafe==1.0
mixer==6.0.1
more-itertools==4.3.0
msgpack==0.5.6
numpy==1.15.4
oauthlib==2.1.0
openapi-codec==1.3.2
packaging==18.0
paho-mqtt==1.3.1
pandas==0.23.4
parso==0.3.1
pbr==4.2.0
pexpect==4.6.0
phonenumbers==8.9.13
pickleshare==0.7.4
Pillow==5.2.0
pkg-resources==0.0.0
pluggy==0.7.1
priority==1.3.0
promise==2.2.1
prompt-toolkit==1.0.15
psycopg2==2.7.5
psycopg2-binary==2.7.5
ptyprocess==0.6.0
purepng==0.2.0
py==1.6.0
pyasn1==0.4.4
pyasn1-modules==0.2.2
pycparser==2.18
pydot==1.2.4
Pygments==2.2.0
PyHamcrest==1.9.0
PyJWT==1.6.4
pykwalify==1.6.1
pyOpenSSL==18.0.0
pyparsing==2.3.0
pytest==3.8.0
pytest-cov==2.6.0
pytest-django==3.4.2
python-box==3.2.1
python-dateutil==2.7.3
python3-openid==3.1.0
pytz==2018.5
PyYAML==3.13
recommonmark==0.4.0
requests==2.19.1
requests-oauthlib==1.0.0
rinoh-typeface-dejavuserif==0.1.1
rinoh-typeface-texgyrecursor==0.1.1
rinoh-typeface-texgyreheros==0.1.1
rinoh-typeface-texgyrepagella==0.1.1
rinohtype==0.3.1
Rx==1.6.1
selenium==3.14.0
service-identity==17.0.0
sexpdata==0.0.3
simplegeneric==0.8.1
simplejson==3.16.0
singledispatch==3.4.0.3
six==1.11.0
snowballstemmer==1.2.1
Sphinx==1.8.2
sphinxcontrib-websupport==1.1.0
sqlparse==0.2.4
stevedore==1.29.0
text-unidecode==1.2
traitlets==4.3.2
Twisted==18.7.0
txaio==18.8.1
uritemplate==3.0.0
urllib3==1.23
wcwidth==0.1.7
Werkzeug==0.14.1
zope.interface==4.5.0
